Extract ingestion job lifecycle boundary
