+++ docs/architecture/CODEBASE-REVIEW-LEDGER.md
@@ -10,6 +10,7 @@ Companion process document:
 
 | Review ID | Date | Scope / Pattern | Status | Findings | Actions Taken | Follow-up | Evidence / Sign-off |
 |---|---|---|---|---|---|---|---|
+| CR-1108 | 2026-06-20 | Ingestion job lifecycle boundary | Hardened | `IngestionJobService` still owned idempotent job creation, queued/failed/retried state transitions, failure-observation recording, simple job reads, replay-context reads, failure listing, and lifecycle metric side effects inline. The public methods were low-complexity, but the service remained the current ingestion maintainability target and mixed facade orchestration with persistence details. | Added `ingestion_job_lifecycle.py` for job/failure mapping, lifecycle dataclasses, idempotent create-or-get persistence, state transitions, failure observation, replay-context reads, and failure listing. Kept `IngestionJobService` public signatures as thin delegates so routers and API contracts remain unchanged. | Continue reducing `ingestion_job_service.py` by cohesive operational responsibilities only when behavior can be pinned with focused tests and measured maintainability movement. | Evidence: [CR-1108-INGESTION-JOB-LIFECYCLE-BOUNDARY.md](./CR-1108-INGESTION-JOB-LIFECYCLE-BOUNDARY.md); focused state-transition tests passed with 4 tests; broader ingestion service unit package passed with 79 tests; scoped Ruff lint/format checks passed; Radon reports `ingestion_job_service.py` improved from `A (25.65)` / 726 SLOC to `A (38.41)` / 584 SLOC and `ingestion_job_lifecycle.py` is `A (40.28)` / 261 SLOC, with all touched service/helper functions A-ranked by cyclomatic complexity. |
 | CR-1107 | 2026-06-20 | App-level supported-surface validation evidence | Hardened | `lotus-core` had strong individual gates and deterministic Docker smoke coverage, but no single app-level evidence command that combined static contracts, supported-feature truth, deterministic synthetic data, live API/calculation proof, and machine-readable validation output. | Added `scripts/certify_lotus_core_app.py` and `make lotus-core-validate`. The command runs OpenAPI, API vocabulary, route-family, source-data product, and domain-product contract checks; verifies supported-feature truth anchors; runs deterministic Docker endpoint smoke over ingestion, event replay/ops, operational reads, support/lineage, integration policy/capabilities, core snapshot, and simulation; and writes JSON evidence under `output/lotus-core-validation/`. Added PR Merge Gate report-only wiring with artifact upload, explicit exit-code evidence, warning emission, and non-blocking workflow semantics while the command remains fail-closed when run directly. | Keep the CI job report-only until lotus-ci-enforcement-governance proves repeated stability, false-positive posture, lane placement, and exception policy. Do not claim standing demo readiness without reviewing the generated evidence for the target runtime. | Evidence: [CR-1107-APP-VALIDATION-EVIDENCE.md](./CR-1107-APP-VALIDATION-EVIDENCE.md); `make lotus-core-validate` passed locally with 66 runtime smoke checks, `make lint`, `make typecheck`, `make warning-gate`, workflow unit tests, and `git diff --check` passed. First remote report-only run exposed missing sibling `lotus-platform` domain-product validator evidence while runtime smoke still passed 66/0; workflow report-only handling now preserves that evidence without failing the PR check. |
 | CR-1106 | 2026-06-17 | Ingestion replay-audit boundary | Hardened | `IngestionJobService` still owned successful replay-audit fingerprint lookup, single-audit lookup, replay-audit row creation, duplicate-blocked metric accounting, and replay-failure metric accounting inline while `ingestion_replay_audits.py` already owned replay-audit response mapping and listing. | Moved replay-audit lookup, single-audit read, audit write, status policy, and metric recording into `ingestion_replay_audits.py`, keeping the service's public methods as thin delegates. Added direct helper tests for successful and missing fingerprint lookup plus persistence and metric routing. | Continue reducing `ingestion_job_service.py` by cohesive operational responsibilities only when behavior can be pinned with focused tests and measured maintainability movement. | Evidence: [CR-1106-INGESTION-REPLAY-AUDIT-BOUNDARY.md](./CR-1106-INGESTION-REPLAY-AUDIT-BOUNDARY.md); focused replay-audit and guardrail tests passed with 21 tests; broader ingestion service unit package passed with 79 tests; scoped Ruff lint/format checks passed; Radon reports `ingestion_job_service.py` improved from `A (22.62)` / 762 SLOC to `A (25.65)` / 726 SLOC and `ingestion_replay_audits.py` is `A (52.41)`. |
 | CR-1105 | 2026-06-17 | Ingestion record-status read-model boundary | Hardened | `IngestionJobService.get_job_record_status` still owned job lookup, failure lookup, malformed request-payload fallback, replayable-record extraction, failed-record aggregation, and response DTO assembly inline. The public method was A-ranked, but the service remained the current ingestion maintainability target and mixed read-model assembly with ingestion lifecycle orchestration. | Moved record-status response assembly and database read-model loading into `ingestion_record_status.py`, keeping `IngestionJobService.get_job_record_status` as a thin public orchestration method. Split record-key extraction helpers so the expanded helper module remains fully A-ranked by cyclomatic complexity. Added direct unit coverage for response assembly, malformed payload fallback, missing-job handling, and loader behavior. | Continue reducing `ingestion_job_service.py` by cohesive operational read responsibilities only when behavior can be pinned with focused tests and measured maintainability movement. | Evidence: focused record-status and ingestion guardrail tests passed with 25 tests; scoped Ruff lint and format checks passed; Radon reports `ingestion_job_service.py` improved from `A (21.58)` / 786 SLOC to `A (22.62)` / 762 SLOC and `ingestion_record_status.py` is `A (48.38)`, with every function A-ranked by cyclomatic complexity. |
