+++ quality/quality_scorecard.md
@@ -1016,3 +1016,8 @@ Local evidence captured on 2026-06-05 after CR-1034:
   `ingestion_job_service.py` from `A (22.62)` / 762 SLOC to `A (25.65)` / 726 SLOC while adding
   direct helper coverage for successful/missing fingerprint lookup, audit persistence, completed
   timestamp posture, and duplicate/failure metric routing.
+- CR-1108 split ingestion job lifecycle persistence, failure observation, simple job reads,
+  replay-context reads, failure listing, response mapping, and lifecycle metric accounting out of
+  `IngestionJobService` into `ingestion_job_lifecycle.py`, improving `ingestion_job_service.py`
+  from `A (25.65)` / 726 SLOC to `A (38.41)` / 584 SLOC while keeping the new helper
+  `A (40.28)` / 261 SLOC and preserving public service signatures.
